requirements/base.txt
# Core (always installed)
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

pandas==2.2.3
numpy==2.1.3

PyYAML==6.0.2
python-dotenv==1.0.1

tqdm==4.67.1
tenacity==9.0.0
